Issue #57

+++ src/ResponsabilitesBundle/Controller/MaterielController.php
@@ -19,7 +19,7 @@ class MaterielController extends Controller
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function newTypeAction(Request $request)
 	{
@@ -27,7 +27,7 @@ class MaterielController extends Controller
 		
 		$form = $this->createForm(TypeFormType::class, $type);
 		
-		if ($request->isType('POST') && $form->handleRequest($request)->isValid()) {
+		if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
 			$em = $this->getDoctrine()->getManager();
 			$em->persist($type);
 			$em->flush();
@@ -51,7 +51,7 @@ class MaterielController extends Controller
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function newObjetAction(Request $request)
 	{
@@ -59,7 +59,7 @@ class MaterielController extends Controller
 		
 		$form = $this->createForm(ObjetFormType::class, $objet);
 		
-		if ($request->isType('POST') && $form->handleRequest($request)->isValid()) {
+		if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
 			$em = $this->getDoctrine()->getManager();
 			$em->persist($objet);
 			$em->flush();
@@ -153,7 +153,7 @@ class MaterielController extends Controller
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function editTypeAction(Request $request, $type)
 	{
@@ -169,7 +169,7 @@ class MaterielController extends Controller
 		
 		$form = $this->createForm(TypeFormType::class, $type);
 		
-		if ($request->isType('POST') && $form->handleRequest($request)->isValid()) {
+		if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
 			$em->persist($type);
 			$em->flush();
 			
@@ -193,7 +193,7 @@ class MaterielController extends Controller
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function editObjetAction(Request $request, $objet)
 	{
@@ -209,7 +209,7 @@ class MaterielController extends Controller
 		
 		$form = $this->createForm(ObjetFormType::class, $objet);
 		
-		if ($request->isType('POST') && $form->handleRequest($request)->isValid()) {
+		if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
 			$em->persist($objet);
 			$em->flush();
 			
@@ -233,7 +233,7 @@ class MaterielController extends Controller
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function deleteTypeAction(Request $request, $type)
 	{
@@ -242,22 +242,31 @@ class MaterielController extends Controller
 
 		$type = $em
 			->getRepository('ResponsabilitesBundle:TypeObjet')
-			->find($type_id);
+			->getWithObjetsById($type_id);
 
 		if (null === $type)
 			throw $this->createNotFoundException("Le type d'objet n°".$type_id." n'existe pas.");
+
+		if (\count($type->getObjets()) != 0) {
+			$this->addFlash('danger', 'materiel.flash.type.delete_abort');
+			return $this->redirectToRoute(
+				'responsabilites_materiel_view_type',
+				array(
+					"type" => $type_id,
+			));
+		}
 		
 		$em->remove($type);
 		$em->flush();
 
 		$this->addFlash('success', 'materiel.flash.type.delete');
 		
-		return $this->redirectToRoute('responsabilites_materiel_view_type');
+		return $this->redirectToRoute('responsabilites_materiel_view_all_type');
 		
 	}
 
 	/**
-	 * @Security("has_role('ROLE_BUDGET')")
+	 * @Security("has_role('ROLE_MATERIEL')")
 	 */
 	public function deleteObjetAction(Request $request, $objet)
 	{
